finished domain layer and app layer for learning paths, sections and units

+++ services/learning-paths/src/app/sections/commands/remove.command.ts
@@ -1,31 +1,22 @@
 import { ICommandHandler, SectionNotFoundException } from '@/app/common';
 import { ILearningPathRepository } from '@/domain/learning-paths/interfaces';
-import { LearningPathId } from '@/domain/learning-paths/value-objects';
 import { ISectionRepository } from '@/domain/sections/interfaces';
 import { LearningPathNotFoundException } from '@app/common';
 import { SectionId } from '@/domain/sections/value-objects/id.vo';
 
-type RemoveSectionFromLearningPathCommand = {
-	learningPathId: string;
+type RemoveSectionCommand = {
 	sectionId: string;
 };
 
-export class RemoveSectionFromLearningPathHandler
-	implements ICommandHandler<RemoveSectionFromLearningPathCommand, void>
+export class RemoveSectionHandler
+	implements ICommandHandler<RemoveSectionCommand, void>
 {
 	constructor(
 		private readonly learningPathRepository: ILearningPathRepository,
 		private readonly sectionRepository: ISectionRepository,
 	) {}
 
-	async execute(command: RemoveSectionFromLearningPathCommand): Promise<void> {
-		const learningPathId = LearningPathId.create(command.learningPathId);
-		const learningPath = await this.learningPathRepository.load(learningPathId);
-
-		if (!learningPath) {
-			throw new LearningPathNotFoundException(command.learningPathId);
-		}
-
+	async execute(command: RemoveSectionCommand): Promise<void> {
 		const sectionId = SectionId.create(command.sectionId);
 		const section = await this.sectionRepository.load(sectionId);
 
@@ -33,8 +24,18 @@ export class RemoveSectionFromLearningPathHandler
 			throw new SectionNotFoundException(sectionId.value);
 		}
 
-		learningPath.removeSection(sectionId);
+		const learningPath = await this.learningPathRepository.load(
+			section.learningPathId,
+		);
+
+		// never going to happen
+		// only for type safety
+		if (!learningPath) {
+			throw new LearningPathNotFoundException(section.learningPathId.value);
+		}
+
 		section.ensureCanRemove();
+		learningPath.removeSection(sectionId);
 
 		await this.sectionRepository.remove(sectionId);
 		await this.learningPathRepository.save(learningPath);
